requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
boto3==1.18.27
botocore==1.21.27
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-redis==5.0.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-jwt==1.9.0
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
kombu==5.1.0
Markdown==3.3.4
prompt-toolkit==3.0.20
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
